backend/requirements-txt.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pypdf2==3.0.1
spacy==3.7.4
python-docx==1.1.0
pydantic==2.6.3
nltk==3.8.1
scikit-learn==1.4.1.post1
numpy==1.26.4
pandas==2.2.1
